+++ ios/Classes/ConnecterManager.m
@@ -115,6 +115,7 @@ static dispatch_once_t once;
     switch (currentConnMethod) {
         case BLUETOOTH:
             _bleConnecter = nil;
+            _connecter = nil;
             break;
     }
 }